arm64: dts: rockchip: Add preinit to antelao bootargs
There are some issues with the package manager on the linux-based system
caused by the init script having a non-standard name.
Since the package manager is required for our use-case to update the
metis driver without updating the system itself, fix this issue by
adding preinit script to the antelao bootargs in its dts.

Change-Id: I9d2fdc1147d88afbdb3b116a851359ebf20364f6
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>