arm64: dts: rockchip: Add preinit to firefly bootargs
There are some issues with the package manager on the linux-based system
caused by the init script having a non-standard name.
Since the package manager is required for our use-case to update the
metis driver without updating the system itself on the antelao board,
we need to add the package manager feature on the BSP.
However, at the moment, this feature would be broken on the firefly.
To fix the issue we must add the preinit script to the firefly bootargs
in its dts.

Since the original `chosen` of the firefly board is inherited by the
`rk3588-linux.dtsi`, copy it to make the change less important for other
boards and to make it easier to maintain.
The newly added chosen is exactly the same as the original, with
`init=/sbin/preinit` appended at the end.

Change-Id: Ib4bcab6f65b0e80e978bd5d8df0fb8c5573b9226
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>