drivers: cpufreq_interactive: Fix `cpufreq_frequency_table_target`
The function definition was updated in `include/linux/cpufreq.h` in
commit 5cb54af11756 ("cpufreq: Fix setting policy limits when frequency
tables are used"), leading to the following
	error: too few arguments to function
In `drivers/cpufreq/cpufreq_interactive.c` while building.

Fix this by passing policy->min and policy->max, which were originally
hardcoded in the function itself and made into arguments with the
previously cited commit.

Fixes: 5cb54af11756 ("cpufreq: Fix setting policy limits when frequency tables are used")
Change-Id: Ib5c52afe04c8bd74c551eecb1ab234d0b7d64ea1
Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>